Restructuring the api
