+++ Makefile
@@ -241,7 +241,7 @@ docker.exporter: build.exporter Dockerfile.exporter
 
 .PHONY: protos
 protos:
-	protoc -I pkg/ndm-grpc/protos/ pkg/ndm-grpc/protos/ndm.proto --go_out=plugins=grpc:pkg/ndm-grpc/protos/ndm
+	protoc -I . ndm.proto --go_out=plugins=grpc:spec/ndm
 
 .PHONY: deps
 deps: header
