+++ pkg/features/features.go
@@ -41,21 +41,21 @@ const (
 	// blockdevice UUID algorithm mentioned in
 	// https://github.com/openebs/openebs/pull/2666
 	GPTBasedUUID Feature = "GPTBasedUUID"
-	// GRPC feature flag starts the GRPC server which provides functionality to manage block devices
-	GRPCServer Feature = "GRPCServer"
+	// APIService feature flag starts the GRPC server which provides functionality to manage block devices
+	APIService Feature = "APIService"
 )
 
 // supportedFeatures is the list of supported features. This is used while parsing the
 // feature flag given via command line
 var supportedFeatures = []Feature{
 	GPTBasedUUID,
-	GRPCServer,
+	APIService,
 }
 
 // defaultFeatureGates is the default features that will be applied to the application
 var defaultFeatureGates = map[Feature]bool{
 	GPTBasedUUID: false,
-	GRPCServer:   false,
+	APIService:   false,
 }
 
 // featureFlag is a map representing the flag and its state
